Revert to a "working" version.
Stoa is great and all, but Demigods being ported over to it was basically a complete rewrite.
If I ever want DG3 to be a thing, it needs to be this way. DG4 is the future, so I'm mainly doing this
to test out features and see what I want to carry over.